refactor(dashboard): keep hand-written icons in src/icons
Artwork behind a React signature is not UI, and mixing the two makes
src/components/ read as though an svg were a component worth composing.
